requirements.txt
accelerate==1.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.14.1
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.11.0
anyio==4.9.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
banal==1.0.6
beautifulsoup4==4.13.4
bitsandbytes==0.43.1
braceexpand==0.1.7
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dataset==1.6.2
datasets==2.21.0
deepspeed==0.14.4
diffusers==0.29.2
dill==0.3.8
distro==1.9.0
dnspython==2.6.1
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.14.0
flash_attn==2.7.4.post1
fonttools==4.52.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.14.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gradio==4.16.0
gradio_client==0.8.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.1
httpx==0.28.1
huggingface-hub==0.29.1
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
-e git+https://github.com/haotian-liu/LLaVA.git@c121f0432da27facab705978f83c4ada465e46fd#egg=llava
Mako==1.3.9
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.78.1
opencv-python==4.10.0.84
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pdf2image==1.17.0
peft==0.11.1
pillow==10.3.0
platformdirs==4.2.2
portalocker==2.10.1
proto-plus==1.26.1
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.7.2
pydantic_core==2.18.3
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.8.0
pynvml==11.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
rsa==4.9.1
ruff==0.4.6
safetensors==0.4.3
-e git+https://github.com/facebookresearch/segment-anything-2.git@82b026cd5578af78757323ab99a0b5c8dc456cff#egg=SAM_2
scikit-learn==1.2.2
scipy==1.13.1
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.3.1
setproctitle==1.3.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
some-package==0.1
soupsieve==2.7
speedtest-cli==2.1.3
SQLAlchemy==1.4.54
starlette==0.37.2
svgwrite==1.4.3
sympy==1.13.1
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.21.0
tomlkit==0.12.0
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tqdm==4.66.4
transformers==4.47.0
triton==3.1.0
trl==0.13.0
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
wandb==0.16.6
watchfiles==0.22.0
wavedrom==2.0.3.post3
webdataset==0.2.111
websockets==15.0.1
xxhash==3.5.0
yarl==1.9.4
zhipuai==2.1.5.20250106
zipp==3.19.2
